requirements.txt
pandas
numpy
scikit-learn
matplotlib
seaborn
mlflow
jupyter
hydra-core
optuna
joblib
omegaconf